ocr-tesseract/requirements.txt
fastapi~=0.110.2
opencv-python~=4.9.0.80
pdf2image~=1.17.0
pillow~=10.3.0
pydantic_settings~=2.2.1
pytesseract~=0.3.10
python-multipart~=0.0.9
redis~=5.0.4
uvicorn[standard]~=0.29.0
